Alteração nos teste, readme e exclusão do endpoint.

+++ src/main/java/com/testedev/readingfile/utils/Utils.java
@@ -58,9 +58,11 @@ public class Utils {
         saida.mkdirs();
     }
 
-    public static void removerArquivosIn(File[] arquivos) {
+    public static void removerArquivosEntrada(File[] arquivos) {
         Arrays.stream(arquivos).forEach(arquivo -> {
-            arquivo.delete();
+            if(isFormatoValido(arquivo.getName())) {
+                arquivo.delete();
+            }
         });
     }
 }
